Have multiple jobs that run in parallel for faster CI. (#14)
Additionally, changed all but one job to directly run from the Ubuntu image (instead of the added indirection of a docker container).

While the parallel jobs should work in theory, and they do when I push to my fork (e.g. advaitjain#2), on the tensorflow/tflite-micro repo, these jobs are queued for a long time.

We will still merge this PR and then investigate why that is the case.