Enforce the pairing of AllocateTempXxx and DeallocateTemp api in all kernels (#906)
* Clean up GetInput/GetOutput api usage in squeeze, transpose, strided_slice op

Clean up api usage in squeeze, transpose, strided_slice op by replacing
GetInput/GetOutput with AllocateTempXXX/DeallocateTempXXX in micro
context api in squeeze, transpose, strided_slice op

Additionally, squeeze and transpose is using TfLiteTensor in Invoke,  which
is againts the recommendation and unnecessary. Replace them with the usage of
TfLiteEvalTensor.

This is a continuation of pr #894 to improve both readability and safety of getting tensors.

The next step is to clean the GetInput/GetOutput in test_helpers.cc
before enforcing the pair of AllocateTempXXX and DeallocateTemp globally.

BUG=https://b/209453859

* Fix bluepill test: need to explictly use int32_t so that the right templated typeToTfLiteType is chosen and return the right type

* Remove redundant include

* Clean GetInput/GetOutput api usage in test_helpers

* Enforce the pairing of AllocateTemp/DeallocateTemp for all kernel tests

* Fix svdf/full_connected in cmsis_nn variant and conv in xtensa variant

* Fix xtensa

* Fix depthwise_conv in xtensa

* Fix xtensa again

* Fix compilation. Xtensa local test pass

* Fix xtensa/full_connected not showing up locally

* update cmsis_nn svdf

* Fix pooling for hifi5

* hif5 errors seen only on server

* Fix hifi5 again