Quantization specific registration for CMSIS-NN (#977)
* Quantization specific registration for CMSIS-NN

Adds int8 and int16x8 specific registrations for conv kernel.

Person detect example is updated with int8 specific
registration for conv. It will result in a smaller memory footprint
for kernels that support this. For kernels that not support it the
default registration will be used.

Also 16x8 conv unit tests are updated with specific 16x8
registration. For kernels that do not have this it will default to
default registration.

Change-Id: I412cc472e56a0835ed662525dfd7768a920590bb

* Address review comments

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>