Enable -Wmissing-field-initializers and update all registration (#1084)
* Enable -Wmissing-field-initializers and update all registration
references

In https://github.com/tensorflow/tflite-micro/pull/1082/, we disabled
the compiler error as a workaround for a successful upstream sync from
TFLite. This PR re-enables the error after addressing all errors. A new
`RegisterOp` API is added, and all prior TfLiteRegistration struct
initializations are updated to use the new API. At the minimum, this
ensures that such additions in future will only require an update to
RegisterOp if TFLM doesn't use that field.

BUG=b/230507399

* Added new RegisterOpWithFree for LSTM

* Remove LSTM Free

* Changed inline to normal function to save code size

* Changed inline to normal function to save code size

* Replace cmsis_nn static initialization with RegisterOp

* Removed extra code

* Add dropped namespace in hard_swish