Hexagon workflow changes to run the hexagon tests from '/opt' directory instead of '/' directory.
At present, when we run the hexagon tests from the docker container, we run the test from root (/) directory. In the scripts, we then find out the current directory for the script and then jump to tflite-micro/ directory and run the rest of the commands from that directory( like we run a bunch of make commands from tflite-micro).

This PR is aiming to simplify that process and make it more aligned with the google3 build system. In this modified hexagon.yml, we are jumping inside the /opt directory and running the test scripts. Now through the docker command, we can pass in additional parameters to the test scripts so that the script can use those parameters to run the make commands relative to the /opt directory.

A follow up PR will have the hexagon test scripts to contain all the changes that are required to use these parameters.

BUG=http://b/244204467