build(make): fix warning about missing dir (#1725)
Don't use `find` to construct a path to Makefile.inc in a generated directory which is not always present, because `find` prints a warning when the directory is absent. Use `$(wildcard)` for consistency with the surrounding definitions. If the path is nonexistent, `$(wildcard)` returns empty, and the include directive which uses the definition below ignores an empty argument.

BUG=fixes #1644, and part of #1663