Adds default values to MicroResourceVariable (#1750)
Currently we memset MicroResourceVariables buffers to 0 if it is null, and in reset we memset to 0 again.
This change makes it so that we memset to the zero_point instead.

To get the information of the zero_point value while reset, we need to store it in the MicroResourceVariable class.

BUG=[269648474](https://b.corp.google.com/issues/269648474)