Deprecate parameter for number of resource variables (#1738)
Continued work on automating resource variables from python. After removing usage of the parameter, this PR removes it completely from the function call.

In addition, we now print the number of resource variables if it is greater than zero, so the user can use it to get the count when doing C++.

BUG=[251851084](https://b.corp.google.com/issues/251851084)