Refactor ci tests to improve speed. (#1746)
This refactor reduces ci test time for all the tests in ci.yml to 16 to 19 minutes depending on weather. 

So our total time on pre-merge ci is going to be ~22 minutes. See bug for more details on runtimes of our current suite. We can get that down to sub 20 minutes if we can move the hifi5 xtensa test to pot-merge which is part of https://github.com/tensorflow/tflite-micro/pull/1772.

This PR will not pass our branch protection automation since it makes substantial changes to our tests. Additionally, the branch protection rules will need to change once this PR is merged.

BUG=https://issuetracker.google.com/issues/266654467