Fix the project generation for hello_world. (#1795)
This PR is going to fix the project generation CI build. 

We are changing evaluate_test to not use the TFLM unit testing framework. This will allow the binary to be used independent of the TFLM repository which can be helpful when porting TFLM.

At the same time, we would like to continue to be able to have evaluate_test be part of the unit test suite which requires returning a specific string when the test is successful.

BUG=http://b/267222141