Fix the lstm evaluate_test. (#1803)
While importing the code for lstm to google3, we found that the `mnist_lstm/evaluate_test.py ` is failing. This is dues to the fact that a number of initialization operations (including I/O) operations were happening outside of any `setUp` method declaration. Please look at the bug for more details.
BUG=http://b/271479216