Change RuntimeShape to support max size of 6 dimensions. (#1882)
Our upstream sync with TF is broken due to [b/275337654](http://b/275337654) which introduces 6D usage in the Add operator.

RuntimeShape.h is managed separately for TFLM vs TFLite.
Since we do not have dynamic memory allocation, we have to explicitly set the max number of dimensions, as previously it was 5.

* Changing the max size to 6 to support the changes to Add.
BUG=[276954955](https://b.corp.google.com/issues/276954955)