Example of simple TF model using variables. (#1890)
To test the fix for [b/277231654,](http://b/277231654) here we introduce the recipes folder and the first example.
A simple accumulator model which showcases usage of resource variables in a multi-subgraph model.

The test causes a crash as expected without the change for `assign_variable.cc`, and passes after adding it.
BUG=[274635545](https://b.corp.google.com/issues/274635545), [277231654](https://b.corp.google.com/issues/277231654)