add  layer_by_layer debugging tool to makefile (#2299)
Make all the code C++ 11 compatible in order for target to be compatible with Embedded Tool Chains / Build Systems.

make_unique isn't available in C++ 11

file that provides a function to Write a Flatbuffer into a file isn't available either in the embedded build systems , so a function was added to save a Flatbuffer into a file.

BUG=[b/288141725](https://b.corp.google.com/issues/288141725)