Remove dynamic allocation from VP6 Reduce kernel (#2394)
The Vision P6 reduce kernel used a temporary std::vector to calculate which axis should be reduced. This commit replaces that with an array of 4 elements because the number of axis to reduce should already be resolved to less than or equal to the input tensor rank, which is also limited to 4.

This commit also aligns variable naming conventions and types in this file for consistency.

BUG=b/287051090